requirements.txt
accelerate==1.6.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
huggingface-hub==0.30.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.0
pillow==11.0.0
psutil==7.0.0
ptflops==0.7.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sympy==1.13.1
tifffile==2025.3.30
timm==1.0.15
torch-ema==0.3
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
compressai
torch-fidelity